refactor(logs): fold rate cap into PostHogQueue, drop PostHogLogsQueue wrapper (#596)
