+++ PostHogTests/PostHogLogsQueueTest.swift
@@ -36,7 +36,7 @@ final class PostHogLogsQueueTests {
         beforeSend: PostHogBeforeSendLogBlock? = nil,
         reachability: Reachability? = nil,
         disableReachabilityForTesting: Bool = true
-    ) -> (PostHogLogsQueue, PostHogConfig) {
+    ) -> (PostHogQueue<PostHogLogRecord>, PostHogConfig) {
         // Unique project token per test → isolated storage folder.
         let token = "logs_test_\(UUID().uuidString)"
         let config = PostHogConfig(projectToken: token, host: "http://localhost:9001")
@@ -51,7 +51,9 @@ final class PostHogLogsQueueTests {
 
         let storage = PostHogStorage(config)
         let api = PostHogApi(config)
-        let queue = PostHogLogsQueue(config, storage, api, reachability)
+        let resourceAttributes = PostHogLogsOTLP.buildResourceAttributes(config.logs)
+        let endpoint = QueueEndpoint<PostHogLogRecord>.logs(api: api, resourceAttributes: resourceAttributes)
+        let queue = PostHogQueue(config, storage, endpoint, reachability)
         // Start without the periodic timer so tests are deterministic.
         // Reachability is opt-in per test via the parameter.
         queue.start(disableReachabilityForTesting: disableReachabilityForTesting, disableQueueTimerForTesting: true)
@@ -525,6 +527,17 @@ final class PostHogLogsQueueTests {
         }
         await waitUntil { queue.depth == 50 }
         #expect(queue.depth == 50)
+    }
+
+    @Test("events and snapshot endpoints opt out of the rate cap")
+    func nonLogsEndpointsDisableRateCap() {
+        // Regression guard: rate cap is a logs-only opt-in. If a future change
+        // accidentally enables it on events / replay, every record would
+        // start being throttled at the queue's add(_:).
+        let config = PostHogConfig(projectToken: "x", host: "http://localhost:9001")
+        let api = PostHogApi(config)
+        #expect(QueueEndpoint<PostHogEvent>.batch(api: api).rateCapMax(config) == 0)
+        #expect(QueueEndpoint<PostHogEvent>.snapshot(api: api).rateCapMax(config) == 0)
     }
 
     // MARK: - beforeSend chain
